Create README.md
For Gephi Data Visualisation Task

Background

One of the most critical aspects of social media analysis is visualising patterns and trends
regarding networks. Platforms like Brandwatch visualise the narratives through word
clouds, but the network analysis requires additional platforms.
Imagine that a democracy-focused NGO is interested in learning which networks are
responsible for seeding and spreading political hashtags in the upcoming August 9th
General Election. They have extracted the mentions from several hashtags and need more
analysis.

Your Task

Your task is to visualise which nodes are significant and what role they are playing in the
spreading and sharing of propaganda. You'll seek to distinguish the key accounts publishing
and posting the hashtag. You will go to the lengths to explain what/who the key accounts
were and which among them were more/less influential. Which among them were verified
and whatever other findings you can learn? You must analyse the material for relevance (as
Nendo has written in its trend reports, there are often ‘trendjackers’ that include the phrase
but do not speak to the issue's significance but are looking to hijack that trending topic on
Twitter to self-promote). Please point this out.